+++ services/121-service/src/registration/registrations.service.ts
@@ -257,7 +257,7 @@ export class RegistrationsService {
     await this.storePhoneNumberInRegistration(programAnswers, referenceId);
     await this.inclusionScoreService.calculateInclusionScore(referenceId);
     await this.inclusionScoreService.calculatePaymentAmountMultiplier(
-      programId,
+      registration.program,
       referenceId,
     );
   }
@@ -962,6 +962,11 @@ export class RegistrationsService {
         RegistrationStatusTimestampField.deleteDate,
       )
       .addOrderBy(`${RegistrationStatusEnum.deleted}.created`, 'DESC')
+      .addSelect(
+        `${RegistrationStatusEnum.completed}.created`,
+        RegistrationStatusTimestampField.completedDate,
+      )
+      .addOrderBy(`${RegistrationStatusEnum.completed}.created`, 'DESC')
       .leftJoin(
         RegistrationStatusChangeEntity,
         RegistrationStatusEnum.startedRegistration,
@@ -1221,7 +1226,7 @@ export class RegistrationsService {
     );
     const calculatedRegistration =
       await this.inclusionScoreService.calculatePaymentAmountMultiplier(
-        registration.program.id,
+        registration.program,
         referenceId,
       );
     if (calculatedRegistration) {
@@ -1269,17 +1274,15 @@ export class RegistrationsService {
   }
 
   public async updateRegistrationStatusBatch(
-    programId: number,
     referenceIdsDto: ReferenceIdsDto,
     registrationStatus: RegistrationStatusEnum,
     message?: string,
     messageContentType?: MessageContentType,
   ): Promise<void> {
-    const program = await this.findProgramOrThrow(programId);
     const errors = [];
     for (const referenceId of referenceIdsDto.referenceIds) {
       const registrationToUpdate = await this.registrationRepository.findOne({
-        where: { referenceId: referenceId, programId: programId },
+        where: { referenceId: referenceId },
       });
       if (!registrationToUpdate) {
         errors.push(`Registration '${referenceId}' is not found`);
@@ -1295,12 +1298,21 @@ export class RegistrationsService {
       }
     }
     if (errors.length === 0) {
+      let programId;
+      let program;
       for (const referenceId of referenceIdsDto.referenceIds) {
         const updatedRegistration = await this.setRegistrationStatus(
           referenceId,
           registrationStatus,
         );
         if (message) {
+          if (updatedRegistration.programId !== programId) {
+            programId = updatedRegistration.programId;
+            // avoid a query per PA if not necessary
+            program = await this.programRepository.findOne({
+              where: { id: programId },
+            });
+          }
           const tryWhatsappFirst =
             registrationStatus === RegistrationStatusEnum.invited
               ? program.tryWhatsAppFirst
@@ -1523,13 +1535,9 @@ export class RegistrationsService {
     return await this.registrationRepository.save(updatedRegistration);
   }
 
-  public async deleteBatch(
-    programId: number,
-    referenceIdsDto: ReferenceIdsDto,
-  ): Promise<void> {
+  public async deleteBatch(referenceIdsDto: ReferenceIdsDto): Promise<void> {
     // Do this first, so that error is already thrown if a PA cannot be changed to deleted, before removing any data below
     await this.updateRegistrationStatusBatch(
-      programId,
       referenceIdsDto,
       RegistrationStatusEnum.deleted,
     );