+++ services/121-service/src/registration/services/bulk-import.service.ts
@@ -349,7 +349,7 @@ export class BulkImportService {
       }
       if (program.paymentAmountMultiplierFormula) {
         await this.inclusionScoreService.calculatePaymentAmountMultiplier(
-          program.id,
+          program,
           registration.referenceId,
         );
       }