+++ services/121-service/test/helpers/utility.helper.ts
@@ -1,6 +1,7 @@
-import crypto from 'crypto';
 import * as request from 'supertest';
 import { DEBUG } from '../../src/config';
+import { SeedScript } from '../../src/scripts/seed-script.enum';
+import { CookieNames } from '../../src/shared/enum/cookie.enums';
 
 export function getHostname(): string {
   return 'http://localhost:3000/api';
@@ -10,41 +11,36 @@ export function getServer(): request.SuperAgentTest {
   return request.agent(getHostname());
 }
 
-export function getIsDebug(): boolean {
-  return DEBUG;
+export const itSkipIfDebug = DEBUG ? it.skip : it;
+
+export function resetDB(seedScript: SeedScript): Promise<request.Response> {
+  return getServer()
+    .post('/scripts/reset')
+    .query({
+      script: seedScript,
+    })
+    .send({
+      secret: process.env.RESET_SECRET,
+    });
 }
 
-export async function resetDB(script: string): Promise<void> {
-  const server = getServer();
-  const resetBody = {
-    secret: process.env.RESET_SECRET,
-  };
-  await server.post('/scripts/reset').query({ script: script }).send(resetBody);
-}
-
-export async function login(): Promise<request.Response> {
-  const body = {
+export function loginAsAdmin(): Promise<request.Response> {
+  return getServer().post(`/user/login`).send({
     username: process.env.USERCONFIG_121_SERVICE_EMAIL_ADMIN,
     password: process.env.USERCONFIG_121_SERVICE_PASSWORD_ADMIN,
-  };
-  const server = getServer();
-  return await server.post(`/user/login`).send(body);
+  });
 }
 
-export function createEspoSignature(
-  payload: any,
-  secret: string,
-  webhookId: string,
-): string {
-  const stringifiedBody = JSON.stringify(payload);
-  const hmac = crypto.createHmac('sha256', secret).update(stringifiedBody);
-  const hmacString = hmac.digest().toString('binary');
-  const concatString = webhookId + ':' + hmacString;
-  const base64encodedString = encodeBase64(concatString);
+export async function getAccessToken(): Promise<string> {
+  const login = await loginAsAdmin();
+  const cookies = login.get('Set-Cookie');
+  const accessToken = cookies
+    .find((cookie: string) => cookie.startsWith(CookieNames.general))
+    .split(';')[0];
 
-  return base64encodedString;
+  return accessToken;
 }
 
-function encodeBase64(data): string {
-  return Buffer.from(data, 'binary').toString('base64');
+export async function waitFor(timeInMs: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, timeInMs));
 }