+++ services/121-service/test/payment/payments.test.ts
@@ -1,5 +1,7 @@
 import { HttpStatus } from '@nestjs/common';
+import { FspName } from '../../src/fsp/enum/fsp-name.enum';
 import { IntersolveJumboResultCode } from '../../src/payments/fsp-integration/intersolve-jumbo/enum/intersolve-jumbo-result-code.enum';
+import { SeedScript } from '../../src/scripts/seed-script.enum';
 import { ProgramPhase } from '../../src/shared/enum/program-phase.model';
 import { StatusEnum } from '../../src/shared/enum/status.enum';
 import {
@@ -11,70 +13,72 @@ import {
   changePaStatus,
   importRegistrations,
 } from '../helpers/registration.helper';
-import { login, resetDB } from '../helpers/utility.helper';
-
-const seedScript = 'nlrc-multiple';
-const programId = 3;
-const referenceId = '63e62864557597e0d';
-const payment = 1;
-const amount = 22;
-const registration = {
-  referenceId: referenceId,
-  preferredLanguage: 'en',
-  paymentAmountMultiplier: 1,
-  firstName: 'John',
-  lastName: 'Smith',
-  phoneNumber: 14155238886,
-  fspName: 'Intersolve-jumbo-physical',
-  whatsappPhoneNumber: 14155238886,
-  addressStreet: 'Teststraat',
-  addressHouseNumber: '1',
-  addressHouseNumberAddition: '',
-  addressPostalCode: '1234AB',
-  addressCity: 'Stad',
-};
-const timer = (ms) => new Promise((res) => setTimeout(res, ms));
-
-let access_token: string;
+import { getAccessToken, resetDB, waitFor } from '../helpers/utility.helper';
 
 describe('Do payment to 1 PA', () => {
+  const programId = 3;
+  const referenceId = '63e62864557597e0d';
+  const payment = 1;
+  const amount = 22;
+  const registration = {
+    referenceId: referenceId,
+    preferredLanguage: 'en',
+    paymentAmountMultiplier: 1,
+    firstName: 'John',
+    lastName: 'Smith',
+    phoneNumber: '14155238886',
+    fspName: FspName.intersolveJumboPhysical,
+    whatsappPhoneNumber: '14155238886',
+    addressStreet: 'Teststraat',
+    addressHouseNumber: '1',
+    addressHouseNumberAddition: '',
+    addressPostalCode: '1234AB',
+    addressCity: 'Stad',
+  };
+
   describe('with FSP: Intersolve Jumbo physical', () => {
+    let accessToken: string;
+
     beforeEach(async () => {
-      await resetDB(seedScript);
-      const loginResponse = await login();
-      access_token = loginResponse.headers['set-cookie'][0].split(';')[0];
+      await resetDB(SeedScript.nlrcMultiple);
+      accessToken = await getAccessToken();
+
       await changePhase(
         programId,
         ProgramPhase.registrationValidation,
-        access_token,
+        accessToken,
       );
-      await changePhase(programId, ProgramPhase.inclusion, access_token);
-      await changePhase(programId, ProgramPhase.payment, access_token);
+      await changePhase(programId, ProgramPhase.inclusion, accessToken);
+      await changePhase(programId, ProgramPhase.payment, accessToken);
     });
 
     it('should succesfully pay-out', async () => {
-      await importRegistrations(programId, [registration], access_token);
-      await changePaStatus(programId, [referenceId], 'include', access_token);
+      // Arrange
+      await importRegistrations(programId, [registration], accessToken);
+      await changePaStatus(programId, [referenceId], 'include', accessToken);
       const paymentReferenceIds = [referenceId];
+
+      // Act
       const doPaymentResponse = await doPayment(
         programId,
         payment,
         amount,
         paymentReferenceIds,
-        access_token,
+        accessToken,
       );
 
       let getTransactionsBody = [];
       while (getTransactionsBody.length <= 0) {
         getTransactionsBody = (
-          await getTransactions(programId, 1, referenceId, access_token)
+          await getTransactions(programId, 1, referenceId, accessToken)
         ).body;
         if (getTransactionsBody.length > 0) {
           break;
         }
-        await timer(2000);
+        await waitFor(2_000);
       }
 
+      // Assert
       expect(doPaymentResponse.status).toBe(HttpStatus.CREATED);
       expect(doPaymentResponse.text).toBe(String(paymentReferenceIds.length));
       expect(getTransactionsBody[0].status).toBe(StatusEnum.success);
@@ -82,29 +86,33 @@ describe('Do payment to 1 PA', () => {
     });
 
     it('should give error about address', async () => {
+      // Arrange
       registration.addressCity = null;
-      await importRegistrations(programId, [registration], access_token);
-      await changePaStatus(programId, [referenceId], 'include', access_token);
+      await importRegistrations(programId, [registration], accessToken);
+      await changePaStatus(programId, [referenceId], 'include', accessToken);
       const paymentReferenceIds = [referenceId];
+
+      // Act
       const doPaymentResponse = await doPayment(
         programId,
         payment,
         amount,
         paymentReferenceIds,
-        access_token,
+        accessToken,
       );
 
       let getTransactionsBody = [];
       while (getTransactionsBody.length <= 0) {
         getTransactionsBody = (
-          await getTransactions(programId, 1, referenceId, access_token)
+          await getTransactions(programId, 1, referenceId, accessToken)
         ).body;
         if (getTransactionsBody.length > 0) {
           break;
         }
-        await timer(2000);
+        await waitFor(2_000);
       }
 
+      // Assert
       expect(doPaymentResponse.status).toBe(HttpStatus.CREATED);
       expect(doPaymentResponse.text).toBe(String(paymentReferenceIds.length));
       expect(getTransactionsBody[0].status).toBe(StatusEnum.error);
@@ -114,7 +122,11 @@ describe('Do payment to 1 PA', () => {
     });
   });
 
-  describe('with FSP: Intersolve AH digital voucher', () => {});
+  describe.skip('with FSP: Intersolve AH digital voucher', () => {
+    it('should succesfully pay-out', () => {});
+  });
 
-  describe('with FSP: Intersolve Visa V-pay', () => {});
+  describe.skip('with FSP: Intersolve Visa V-pay', () => {
+    it('should succesfully pay-out', () => {});
+  });
 });